feat: Basic setup for working with packer.nvim
After what felt like an eternity of trying to figure out how to
get `packer.nvim` in as an optional package, I gave up.

What we have no is `packer` installed in the way that their README
suggests. A slight difference that I am toying with, is to separate each
plugins as it's own module under the plugin folder. The
`/plugins/init.lua` is what sets everything up and it does so by
iterating through every file in the plugin folder and loading them.

To make things somewhat easier, I created and autocommand that will
reload all configurations whenever a write is made to a file that exists
in the folder.

So far, things kinda work. Next time, I want to add a plugin, I simply
write add a new plugin spec file in the plugins folder. It seems to work
for now, so... fuck it.